clarify "asterisk usage as multiply sign" clause
The original clause assumed that the asterisk is always used in place of
the multiplication sign, and considers its usage almost always an error.
However, there are possible uses of the asterisk in math that are un-
related to multiplication, making this clause confusing.

This change makes it clear that using asterisk in place of a proper mul-
tiplication mark is discouraged, without prohibiting other uses. This
hopefully makes things more clear.